Update and rename lecture1.py to hw1.py

+++ hw1.py
@@ -1,6 +1,6 @@
-a=int(input())
-b=int(input())
-x=int(input())
+a=float(input())
+b=float(input())
+x=float(input())
 import math
 x=(x*math.pi)/180
 print(math.sqrt(a**2+b**2-2*a*b*math.cos(x)))
